pyproject.toml
[tool.poetry]
name = "drf-auth2"
version = "0.1.0"
description = ""
authors = ["Bhagirath Bhatt <budbhatt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
djangorestframework = "^3.12.4"
django-environ = "^0.4.5"
psycopg2-binary = "^2.8.6"
djangorestframework-simplejwt = "^4.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
